refactor(conversion): 입력값을 객체로 변환하는 작업을 별도의 Converter를 사용하지 않고 생성자를 통해 변환할 수 있도록 수정
생성자가 public으로 열려있는 상황에서 객체를 생성하기 위해 생성자를 써야하는 지 Converter를 써야하는 지 개발자가 헷갈릴 수 있으므로, 객체를 생성하기 위한 모든 활동을 생성자를 통해서만 실행하도록 Converter 자체를 애플리케이션에서 제거
