+++ src/main/java/lotto/vo/LottoBuyingInfo.java
@@ -1,5 +1,7 @@
 package lotto.vo;
 
+import lotto.system.holder.ValidationHolder;
+
 public class LottoBuyingInfo {
 	private final int money;
 	private final int amount;
@@ -7,6 +9,19 @@ public class LottoBuyingInfo {
 	public LottoBuyingInfo(int money, int amount) {
 		this.money = money;
 		this.amount = amount;
+	}
+
+	public LottoBuyingInfo(String input) {
+		ValidationHolder.validate(input, LottoBuyingInfo.class);
+		int money = Integer.parseInt(removeCommaFrom(input));
+		ValidationHolder.validate(money, LottoBuyingInfo.class);
+
+		this.money = money;
+		this.amount = money / 1000;
+	}
+
+	private static String removeCommaFrom(String target) {
+		return target.replaceAll(",", "");
 	}
 
 	public int getMoney() {
