+++ src/test/java/lotto/LottoTest.java
@@ -15,7 +15,6 @@ class LottoTest {
     @BeforeEach
     void setup() {
         LottoApplication.initializeValidators();
-        LottoApplication.initializeConverters();
     }
 
     @AfterEach
