+++ src/test/java/lotto/system/validator/ValidatorTest.java
@@ -23,7 +23,6 @@ class ValidatorTest {
     @BeforeEach
     void setup() {
         LottoApplication.initializeValidators();
-        LottoApplication.initializeConverters();
     }
 
     @AfterEach
